+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8N5W20AR00T2VBTN905K81V"
+        "01K8NEB44R7KPCBGSP43YV7MHV"
     )
     r"""The unique identifier of the data source"""