+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -63,5 +63,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-28T10:29:48.802Z")
+    created: Optional[datetime] = parse_datetime("2025-10-28T12:57:50.819Z")
     r"""The date and time the resource was created"""