+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-28T10:29:45.003Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-28T12:57:46.995Z")
     r"""The date and time the resource was last updated"""