+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8N5W2061VGPBBSC9HCJXNQ5"
+        "01K8NEB44N0PFRC9AWXPBK1598"
     )
     r"""The unique identifier of the data source"""