+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8N5W20EGN7XP2D7Q48N6HJ5"
+        "01K8NEB44V135PMC8F2EDBPWPJ"
     )
     r"""The unique identifier of the data source"""